refactor: modularize sonar-idl parser
