+++ crates/sonar-idl/src/models.rs
@@ -1,5 +1,5 @@
 use serde::de::Error;
-use serde::{Deserialize, Serialize};
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use serde_json::Value as JsonValue;
 
 use crate::discriminator::{sighash, to_snake_case};
@@ -15,8 +15,10 @@ pub enum RawAnchorIdl {
 impl RawAnchorIdl {
     pub fn convert(self, program_address: &str) -> Idl {
         match self {
-            RawAnchorIdl::Current(idl) => idl,
-            RawAnchorIdl::Legacy(legacy) => legacy.into_idl(program_address),
+            RawAnchorIdl::Current(idl) => idl.normalize(program_address),
+            RawAnchorIdl::Legacy(legacy) => {
+                legacy.into_idl(program_address).normalize(program_address)
+            }
         }
     }
 }
@@ -46,30 +48,6 @@ impl LegacyIdl {
             types.extend(accounts);
         }
 
-        let instructions: Vec<IdlInstruction> = self
-            .instructions
-            .into_iter()
-            .map(|mut inst| {
-                if inst.discriminator.is_none() {
-                    let snake_name = to_snake_case(&inst.name);
-                    inst.discriminator = Some(sighash("global", &snake_name).to_vec());
-                }
-                inst
-            })
-            .collect();
-
-        let events: Option<Vec<IdlEvent>> = self.events.map(|events| {
-            events
-                .into_iter()
-                .map(|mut event| {
-                    if event.discriminator.is_none() {
-                        event.discriminator = Some(sighash("event", &event.name).to_vec());
-                    }
-                    event
-                })
-                .collect()
-        });
-
         let metadata = self.metadata.unwrap_or_else(|| IdlMetadata {
             name: self.name.clone(),
             version: self.version.clone(),
@@ -80,11 +58,42 @@ impl LegacyIdl {
         Idl {
             address: address.to_string(),
             metadata,
-            instructions,
+            instructions: self.instructions,
             types: if types.is_empty() { None } else { Some(types) },
-            events,
+            events: self.events,
         }
     }
+}
+
+impl Idl {
+    fn normalize(mut self, fallback_address: &str) -> Self {
+        if self.address.is_empty() {
+            self.address = fallback_address.to_string();
+        }
+
+        self.instructions = self.instructions.into_iter().map(normalize_instruction).collect();
+
+        self.events = self.events.map(|events| events.into_iter().map(normalize_event).collect());
+
+        self
+    }
+}
+
+fn normalize_instruction(mut instruction: IdlInstruction) -> IdlInstruction {
+    if instruction.discriminator.is_none() {
+        let snake_name = to_snake_case(&instruction.name);
+        instruction.discriminator = Some(sighash("global", &snake_name).to_vec());
+    }
+
+    instruction
+}
+
+fn normalize_event(mut event: IdlEvent) -> IdlEvent {
+    if event.discriminator.is_none() {
+        event.discriminator = Some(sighash("event", &event.name).to_vec());
+    }
+
+    event
 }
 
 /// Complete IDL structure including types for full type resolution.
@@ -121,8 +130,8 @@ pub struct IdlEvent {
     pub name: String,
     #[serde(default)]
     pub discriminator: Option<Vec<u8>>,
-    #[serde(default)]
-    pub fields: Option<Vec<IdlField>>,
+    #[serde(default, deserialize_with = "deserialize_optional_idl_fields")]
+    pub fields: Option<IdlFields>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -165,8 +174,27 @@ pub enum IdlType {
     Simple(String),
     Vec { vec: Box<IdlType> },
     Option { option: Box<IdlType> },
-    Array { array: [JsonValue; 2] },
+    Array { array: IdlArrayType },
     Defined { defined: DefinedType },
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[serde(from = "(IdlType, usize)", into = "(IdlType, usize)")]
+pub struct IdlArrayType {
+    pub element_type: Box<IdlType>,
+    pub length: usize,
+}
+
+impl From<(IdlType, usize)> for IdlArrayType {
+    fn from((element_type, length): (IdlType, usize)) -> Self {
+        Self { element_type: Box::new(element_type), length }
+    }
+}
+
+impl From<IdlArrayType> for (IdlType, usize) {
+    fn from(value: IdlArrayType) -> Self {
+        (*value.element_type, value.length)
+    }
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
@@ -197,24 +225,58 @@ pub struct IdlTypeDefinition {
     pub type_: IdlTypeDefinitionBody,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct IdlTypeDefinitionBody {
-    pub kind: String,
+    pub kind: IdlTypeDefinitionKind,
     #[serde(default, deserialize_with = "deserialize_optional_idl_fields")]
     pub fields: Option<IdlFields>,
     #[serde(default)]
     pub variants: Option<Vec<IdlEnumVariant>>,
 }
 
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum IdlTypeDefinitionKind {
+    Struct,
+    Enum,
+    Other(String),
+}
+
+impl Serialize for IdlTypeDefinitionKind {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        serializer.serialize_str(match self {
+            IdlTypeDefinitionKind::Struct => "struct",
+            IdlTypeDefinitionKind::Enum => "enum",
+            IdlTypeDefinitionKind::Other(value) => value,
+        })
+    }
+}
+
+impl<'de> Deserialize<'de> for IdlTypeDefinitionKind {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let value = String::deserialize(deserializer)?;
+        Ok(match value.as_str() {
+            "struct" => IdlTypeDefinitionKind::Struct,
+            "enum" => IdlTypeDefinitionKind::Enum,
+            _ => IdlTypeDefinitionKind::Other(value),
+        })
+    }
+}
+
 /// Fields for IDL types — named (regular structs) or tuple (tuple structs).
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(untagged)]
 pub enum IdlFields {
     Named(Vec<IdlField>),
-    Tuple(Vec<String>),
+    Tuple(Vec<IdlType>),
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct IdlField {
     pub name: String,
     #[serde(rename = "type")]
@@ -243,7 +305,7 @@ fn deserialize_idl_fields_(value: serde_json::Value) -> Result<Option<IdlFields>
                     Err(e) => Err(format!("Failed to parse named fields: {}", e)),
                 }
             } else {
-                match serde_json::from_value::<Vec<String>>(serde_json::Value::Array(arr)) {
+                match serde_json::from_value::<Vec<IdlType>>(serde_json::Value::Array(arr)) {
                     Ok(types) => Ok(Some(IdlFields::Tuple(types))),
                     Err(e) => Err(format!("Failed to parse tuple fields: {}", e)),
                 }
@@ -253,11 +315,11 @@ fn deserialize_idl_fields_(value: serde_json::Value) -> Result<Option<IdlFields>
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct IdlEnumVariant {
     pub name: String,
-    #[serde(default)]
-    pub fields: Option<JsonValue>,
+    #[serde(default, deserialize_with = "deserialize_optional_idl_fields")]
+    pub fields: Option<IdlFields>,
 }
 
 #[cfg(test)]
@@ -339,6 +401,75 @@ mod tests {
     }
 
     #[test]
+    fn current_idl_instruction_gets_auto_discriminator_on_convert() {
+        let raw: RawAnchorIdl = serde_json::from_str(
+            r#"{
+                "address": "11111111111111111111111111111111",
+                "metadata": { "name": "current_program", "version": "0.1.0", "spec": "0.1.0" },
+                "instructions": [{
+                    "name": "doSomething",
+                    "accounts": [],
+                    "args": []
+                }]
+            }"#,
+        )
+        .unwrap();
+
+        let idl = raw.convert("11111111111111111111111111111111");
+        let instruction = &idl.instructions[0];
+
+        assert_eq!(instruction.discriminator, Some(sighash("global", "do_something").to_vec()));
+    }
+
+    #[test]
+    fn current_idl_event_gets_auto_discriminator_on_convert() {
+        let raw: RawAnchorIdl = serde_json::from_str(
+            r#"{
+                "address": "11111111111111111111111111111111",
+                "metadata": { "name": "current_program", "version": "0.1.0", "spec": "0.1.0" },
+                "instructions": [],
+                "events": [{
+                    "name": "TransferEvent",
+                    "fields": [{ "name": "amount", "type": "u64" }]
+                }]
+            }"#,
+        )
+        .unwrap();
+
+        let idl = raw.convert("11111111111111111111111111111111");
+        let event = &idl.events.as_ref().unwrap()[0];
+
+        assert_eq!(event.discriminator, Some(sighash("event", "TransferEvent").to_vec()));
+    }
+
+    #[test]
+    fn current_idl_event_fields_support_tuple_types() {
+        let raw: RawAnchorIdl = serde_json::from_str(
+            r#"{
+                "address": "11111111111111111111111111111111",
+                "metadata": { "name": "current_program", "version": "0.1.0", "spec": "0.1.0" },
+                "instructions": [],
+                "events": [{
+                    "name": "PairEvent",
+                    "fields": ["u32", {"option":"u16"}]
+                }]
+            }"#,
+        )
+        .unwrap();
+
+        let idl = raw.convert("11111111111111111111111111111111");
+        let event = &idl.events.as_ref().unwrap()[0];
+
+        assert_eq!(
+            event.fields,
+            Some(IdlFields::Tuple(vec![
+                IdlType::Simple("u32".into()),
+                IdlType::Option { option: Box::new(IdlType::Simple("u16".into())) },
+            ]))
+        );
+    }
+
+    #[test]
     fn parse_legacy_idl_and_convert() {
         let raw: RawAnchorIdl = serde_json::from_str(LEGACY_IDL_JSON).unwrap();
         let idl = raw.convert("11111111111111111111111111111111");
@@ -415,12 +546,82 @@ mod tests {
                 r#"{"defined":"MyStruct"}"#,
                 IdlType::Defined { defined: DefinedType::Simple("MyStruct".into()) },
             ),
+            (
+                r#"{"array":[{"defined":"MyStruct"},3]}"#,
+                IdlType::Array {
+                    array: IdlArrayType {
+                        element_type: Box::new(IdlType::Defined {
+                            defined: DefinedType::Simple("MyStruct".into()),
+                        }),
+                        length: 3,
+                    },
+                },
+            ),
         ];
 
         for (json_str, expected) in &types {
             let parsed: IdlType = serde_json::from_str(json_str).unwrap();
             assert_eq!(&parsed, expected, "failed for {}", json_str);
         }
+    }
+
+    #[test]
+    fn idl_type_definition_kind_serde_roundtrip() {
+        let struct_def: IdlTypeDefinition = serde_json::from_str(
+            r#"{
+                "name": "MyStruct",
+                "type": {
+                    "kind": "struct",
+                    "fields": [{ "name": "value", "type": "u64" }]
+                }
+            }"#,
+        )
+        .unwrap();
+        assert_eq!(struct_def.type_.kind, IdlTypeDefinitionKind::Struct);
+
+        let enum_def: IdlTypeDefinition = serde_json::from_str(
+            r#"{
+                "name": "MyEnum",
+                "type": {
+                    "kind": "enum",
+                    "variants": [{ "name": "Ready" }]
+                }
+            }"#,
+        )
+        .unwrap();
+        assert_eq!(enum_def.type_.kind, IdlTypeDefinitionKind::Enum);
+    }
+
+    #[test]
+    fn parse_current_idl_tuple_fields_with_nested_types() {
+        let parsed: Result<Idl, _> = serde_json::from_str(
+            r#"{
+                "address": "11111111111111111111111111111111",
+                "metadata": { "name": "tuple_types", "version": "0.1.0", "spec": "0.1.0" },
+                "instructions": [],
+                "types": [
+                    {
+                        "name": "Inner",
+                        "type": {
+                            "kind": "struct",
+                            "fields": [{ "name": "amount", "type": "u16" }]
+                        }
+                    },
+                    {
+                        "name": "Wrapper",
+                        "type": {
+                            "kind": "struct",
+                            "fields": [
+                                { "option": "u32" },
+                                { "defined": "Inner" }
+                            ]
+                        }
+                    }
+                ]
+            }"#,
+        );
+
+        assert!(parsed.is_ok(), "tuple fields should support full IDL types: {:?}", parsed.err());
     }
 
     #[test]
